refactor(util): fold the hash helper into util/json.rs
Review question: a one-function util/hash.rs doesn't earn its own file.
Both consumers (action-registry fingerprints, scan-cache keys) use the
canonicalize-then-digest combination, so the digest now lives next to
canonical_json in util/json.rs; util/hash.rs is removed. (The current
shared digest is BLAKE3 for collision resistance — preserved as-is.)
