+++ crates/skardi/src/util/mod.rs
@@ -1,5 +1,4 @@
 //! Small shared utilities with no better home.
 
-pub mod hash;
 pub mod http;
 pub mod json;
